+++ packages/plugin-autocapture-browser/src/observables.ts
@@ -1,3 +1,4 @@
+import { TimestampedEvent } from './helpers';
 import { Observable, consoleObserver, getGlobalScope, merge } from '@amplitude/analytics-core';
 
 /* eslint-disable-next-line no-restricted-globals */
@@ -31,9 +32,23 @@ export const createClickObservable = (
     const handler = (event: MouseEvent | PointerEvent) => {
       observer.next(event);
     };
-    globalScope.document.addEventListener(clickType, handler, { capture: true });
+
+    getGlobalScope()?.document.addEventListener(clickType, handler, { capture: true });
     return () => {
-      globalScope.document.removeEventListener(clickType, handler, { capture: true });
+      getGlobalScope()?.document.removeEventListener(clickType, handler, { capture: true });
+    };
+  });
+};
+
+export const createScrollObservable = (): Observable<Event> => {
+  return new Observable<Event>((observer) => {
+    const handler = (event: Event) => {
+      observer.next(event);
+    };
+
+    getGlobalScope()?.window.addEventListener('scroll', handler);
+    return () => {
+      getGlobalScope()?.window.removeEventListener('scroll', handler);
     };
   });
 };
@@ -55,6 +70,54 @@ const createConsoleErrorObservable = (): Observable<BrowserErrorEvent> => {
   });
 };
 
+// Tracks when a trackedelement is exposed to the viewport
+export const createExposureObservable = (
+  mutationObservable: Observable<TimestampedEvent<MutationRecord[]>>,
+  selectorAllowlist: string[],
+): Observable<Event> => {
+  return new Observable<Event>((observer) => {
+    const globalScope = getGlobalScope();
+
+    if (!globalScope?.IntersectionObserver) {
+      console.log('IntersectionObserver not supported');
+      return () => {
+        return;
+      };
+    }
+
+    const intersectionObserver = new IntersectionObserver(
+      (entries) => {
+        entries.forEach((entry) => {
+          observer.next(entry as unknown as Event);
+        });
+      },
+      {
+        root: null, // viewport
+        rootMargin: '0px', // start exactly at the viewport edge
+        threshold: 1.0, // trigger when 100% of the element is visible
+      },
+    );
+
+    // Observe initial elements
+    const selectorString = selectorAllowlist.join(',');
+    /* istanbul ignore next */
+    const initialElements = globalScope?.document.querySelectorAll(selectorString) ?? [];
+    initialElements.forEach((element) => {
+      intersectionObserver.observe(element);
+    });
+
+    // Use mutation observable to observe new elements
+    mutationObservable.subscribe(({ event }) =>
+      event.forEach(({ addedNodes }) =>
+        addedNodes.forEach((node) => node instanceof Element && intersectionObserver.observe(node)),
+      ),
+    );
+
+    return () => {
+      intersectionObserver.disconnect();
+    };
+  });
+};
 const createUnhandledErrorObservable = (): Observable<BrowserErrorEvent> => {
   return new Observable<BrowserErrorEvent>((observer) => {
     const handler = (event: Event) => {