+++ packages/plugin-web-attribution-browser/test/web-attribution.test.ts
@@ -2,8 +2,7 @@ import { createInstance } from '@amplitude/analytics-browser';
 import { BASE_CAMPAIGN, CampaignParser, CookieStorage, FetchTransport } from '@amplitude/analytics-client-common';
 import { webAttributionPlugin } from '../src/web-attribution';
 import * as helpers from '@amplitude/analytics-client-common';
-import { BrowserConfig, LogLevel } from '@amplitude/analytics-core';
-import { Logger, UUID } from '@amplitude/analytics-core';
+import { BrowserConfig, LogLevel, Logger, UUID } from '@amplitude/analytics-core';
 
 describe('webAttributionPlugin', () => {
   const mockConfig: BrowserConfig = {